refactor: rename files to match guidelines
