+++ src/api/vehicles/index.ts
@@ -1,7 +1,7 @@
 import Hapi from '@hapi/hapi';
 import {IVehiclesService} from '../../service/interface';
 import {ServiceJourneyVehicleQueryVariables} from '../../service/types';
-import {getUrlParam} from '../../utils/getUrlParam';
+import {getUrlParam} from '../../utils/get-url-param';
 import {
   getVehiclesRequest,
   postServiceJourneySubscriptionRequest,
