+++ src/service/types.ts
@@ -1,6 +1,4 @@
 import {Location, QueryMode} from '@entur/sdk';
-import {boomify} from '@hapi/boom';
-import {FetchError, Response} from 'node-fetch';
 import {
   Mode,
   PointsOnLink,
@@ -10,8 +8,6 @@ import {FormFactor} from '../graphql/mobility/mobility-types_v2';
 import * as Types from '../graphql/vehicles/vehicles-types_v1';
 import {CursoredQuery} from './cursored';
 import {GetServiceJourneyVehicleQuery} from './impl/vehicles/vehicles-gql/vehicles.graphql-gen';
-
-const STATUS_CODES_TO_FORWARD = [429];
 
 export interface Coordinates {
   latitude: number;
@@ -214,41 +210,3 @@ export type VippsCustomTokenRequest = {
   state: string;
   nonce: string;
 };
-
-export class APIError extends Error {
-  public statusCode?: number = 500;
-
-  constructor(error: any) {
-    super();
-    if (error instanceof FetchError) {
-      switch (error.code) {
-        case 'ETIMEDOUT':
-        case 'EPIPE':
-        case 'ECONNRESET':
-        case 'ECONNREFUSED':
-        case 'ENOTFOUND':
-          this.message = 'Upstream service temporarily unavailable';
-          this.statusCode = 503;
-      }
-    }
-    const networkErrorResponse = getNetworkErrorResponse(error);
-    if (
-      networkErrorResponse &&
-      STATUS_CODES_TO_FORWARD.includes(networkErrorResponse.status)
-    ) {
-      this.message = networkErrorResponse.statusText;
-      this.statusCode = networkErrorResponse.status;
-    }
-
-    return boomify(this, {
-      statusCode: this.statusCode,
-      message: error.message,
-    });
-  }
-}
-
-function getNetworkErrorResponse(error: any): Response | undefined {
-  if (error?.networkError?.response instanceof Response) {
-    return error.networkError.response;
-  }
-}
